added route for deleting comments and fetch request is component

+++ server/routes/ticket/deleteComment.js
@@ -1,17 +1,51 @@
 const express = require("express");
 const Project = require("../../models/project");
 const Ticket = require("../../models/ticket");
-const TicketUser = require("../../models/ticketUser");
 const ProjectUser = require("../../models/projectUser");
-const comment = require('../../models/comment');
+const Comment = require("../../models/comment");
 const auth = require("../../verifyJWT");
+const comment = require("../../models/comment");
 
 const deleteComment = express.Router();
 
+const isCreator = async (commentId, userId) => {
+  const comment = await Comment.findById(commentId);
+  return comment.creator === userId;
+};
+
+const isManager = async (projectId, userId) => {
+  const projectUser = await ProjectUser.findOne({
+    user_id: userId,
+    project_id: projectId,
+  });
+};
+
+const isAdmin = (teamRole) => teamRole === "admin";
+
 deleteComment.route("/deleteComment").post(auth.verifyJWT, async (req, res) => {
+  const commentId = req.body.comment;
+  const projectId = req.body.project;
+  const ticketId = req.body.ticket;
+  const teamRole = req.user.team.role;
+  const userId = req.user.id;
   try {
-    const commentId = req.body.comment;
+    if (
+      !isAdmin(teamRole) &&
+      !isCreator(commentId, userId) &&
+      !isManager(projectId, userId)
+    ) {
+      return res.json({
+        failed: true,
+        message: "You do not have permission to delete this comment",
+      });
+    }
+
+    const ticket = await Ticket.findById(ticketId);
+    ticket.comments = ticket.comments.filter(e => e!== commentId);
+    await ticket.save();
+
     await comment.deleteOne({_id: commentId});
+
   } catch (err) {
     console.log(err);
     res.json({
@@ -20,4 +54,3 @@ deleteComment.route("/deleteComment").post(auth.verifyJWT, async (req, res) => {
     });
   }
 });
-
